# src/main/java/ru/kolesnikov/simplechat/controller/dto/UserDTOAuth.java
package ru.kolesnikov.simplechat.controller.dto;

import lombok.Getter;
import lombok.NonNull;
import lombok.RequiredArgsConstructor;

@RequiredArgsConstructor
@Getter
public class UserDTOAuth {
    @NonNull
    private final String login;
    @NonNull
    private final String password;
}